commit end day 2/4 by hoadang
