+++ server/src/app/controllers/attendanceController.go
@@ -72,14 +72,25 @@ func getAttendancesByInternID(c *gin.Context, idTrainee bson.ObjectId) []models.
 	return attens
 }
 
-func GetTraineeAttendances(c *gin.Context) {
+func GetInternAttendances(c *gin.Context) {
 	idTrainee := bson.ObjectIdHex(c.Param("id"))
+	database := c.MustGet("db").(*mgo.Database)
+	intern := models.Intern{}
+	course := models.Course{}
+	database.C(models.CollectionIntern).Find(bson.M{"_id": idTrainee, "IsDeleted": false}).One(&intern)
+	database.C(models.CollectionCourse).Find(bson.M{"_id": intern.CourseID, "IsDeleted": false}).One(&course)
 	attens := getAttendancesByInternID(c, idTrainee)
-	c.JSON(http.StatusOK, attens)
+	resp := Attendance{}
+	resp.Name = intern.Name
+	resp.StartDate = course.StartDate
+	resp.EndDate = course.EndDate
+	resp.Course = course.CourseName
+	resp.Attendances = attens
+	c.JSON(http.StatusOK, resp)
 
 }
 
-type AttendanceMentor struct {
+type Attendance struct {
 	Id          string
 	Name        string
 	Course      string
@@ -88,7 +99,7 @@ type AttendanceMentor struct {
 	Attendances []models.Attendance
 }
 
-func getAttendancesByMentorId(c *gin.Context, idMentor bson.ObjectId) []AttendanceMentor {
+func getAttendancesByMentorId(c *gin.Context, idMentor bson.ObjectId) []Attendance {
 	database := c.MustGet("db").(*mgo.Database)
 
 	courses := []models.Course{}
@@ -98,7 +109,7 @@ func getAttendancesByMentorId(c *gin.Context, idMentor bson.ObjectId) []Attendan
 	}
 
 	//courseTrainee := []models.Intern
-	resp := []AttendanceMentor{}
+	resp := []Attendance{}
 	for _, course := range courses {
 		interns := []models.Intern{}
 		err = database.C(models.CollectionIntern).Find(bson.M{"CourseID": course.ID, "IsDeleted": false}).All(&interns)
@@ -106,7 +117,7 @@ func getAttendancesByMentorId(c *gin.Context, idMentor bson.ObjectId) []Attendan
 			return nil
 		}
 		for _, intern := range interns {
-			resp = append(resp, AttendanceMentor{
+			resp = append(resp, Attendance{
 				Id:          intern.ID.Hex(),
 				Name:        intern.Name,
 				Course:      course.CourseName,
@@ -132,7 +143,7 @@ func GetAttendancesByMentor(c *gin.Context) {
 
 type AttendanceSupervisor struct {
 	Name                string
-	AttendancesByMentor []AttendanceMentor
+	AttendancesByMentor []Attendance
 }
 
 func GetAttendancesBySupervisor(c *gin.Context) {
@@ -169,7 +180,7 @@ func getDailyAttendancesByInternID(c *gin.Context, idTrainee bson.ObjectId, date
 	return atten
 }
 
-func getDailyAttendancesByMentorId(c *gin.Context, idMentor bson.ObjectId, date time.Time) []AttendanceMentor {
+func getDailyAttendancesByMentorId(c *gin.Context, idMentor bson.ObjectId, date time.Time) []Attendance {
 	database := c.MustGet("db").(*mgo.Database)
 
 	courses := []models.Course{}
@@ -188,12 +199,12 @@ func getDailyAttendancesByMentorId(c *gin.Context, idMentor bson.ObjectId, date 
 		courseTrainee[course.CourseName] = interns
 	}
 
-	resp := []AttendanceMentor{}
+	resp := []Attendance{}
 
 	for course, trainees := range courseTrainee {
 		for _, trainee := range trainees {
 			atten := getDailyAttendancesByInternID(c, trainee.ID, date)
-			data := AttendanceMentor{Id: trainee.ID.Hex(), Name: trainee.Name, Course: course, Attendances: atten}
+			data := Attendance{Id: trainee.ID.Hex(), Name: trainee.Name, Course: course, Attendances: atten}
 			resp = append(resp, data)
 		}
 	}
