+++ server/src/app/main.go
@@ -61,7 +61,7 @@ func setupRouter() *gin.Engine {
 	//attendance
 	r.GET("/attendance/:id", controllers.GetAttendance)
 	r.GET("/attendances", controllers.GetListAttendances)
-	r.GET("/attendance/:id/trainee", controllers.GetTraineeAttendances)
+	r.GET("/attendance/:id/intern", controllers.GetInternAttendances)
 	r.GET("/attendance/:id/mentor", controllers.GetAttendancesByMentor)
 	r.GET("/attendance/:id/supervisor", controllers.GetAttendancesBySupervisor)
 	r.GET("/attendance/:id/mentor/daily", controllers.GetDailyAttendanceByMentor)
